feat: errors

+++ controllers/cards.js
@@ -1,42 +1,55 @@
+const { isValidObjectId, Error } = require('mongoose');
 const Card = require('../models/card');
-const { handleCardErrors } = require('../utils/errors');
+const { handleCardError } = require('../utils/errors');
 
 module.exports.createCard = (req, res) => {
   const { name, link } = req.body;
   const owner = req.user._id;
   Card.create({ name, link, owner })
     .then((card) => res.send(card))
-    .catch((err) => handleCardErrors(err, res));
+    .catch((err) => handleCardError(err, res));
 };
 
 module.exports.getAllCards = (req, res) => {
   Card.find({})
     .then((cards) => res.send(cards))
-    .catch((err) => handleCardErrors(err, res));
+    .catch((err) => handleCardError(err, res));
 };
 
 module.exports.deleteCard = (req, res) => {
+  if (!isValidObjectId(req.params.id)) {
+    handleCardError(new Error.CastError(), res);
+    return;
+  }
   Card.findByIdAndRemove(req.params.id)
     .then((card) => res.send(card))
-    .catch((err) => handleCardErrors(err, res));
+    .catch((err) => handleCardError(err, res));
 };
 
 module.exports.likeCard = (req, res) => {
+  if (!isValidObjectId(req.params.id)) {
+    handleCardError(new Error.CastError(), res);
+    return;
+  }
   Card.findByIdAndUpdate(
     req.params.id,
     { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
     { new: true },
   )
     .then((card) => res.send(card))
-    .catch((err) => handleCardErrors(err, res));
+    .catch((err) => handleCardError(err, res));
 };
 
 module.exports.dislikeCard = (req, res) => {
+  if (!isValidObjectId(req.params.id)) {
+    handleCardError(new Error.CastError(), res);
+    return;
+  }
   Card.findByIdAndUpdate(
     req.params.id,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
     { new: true },
   )
     .then((card) => res.send(card))
-    .catch((err) => handleCardErrors(err, res));
+    .catch((err) => handleCardError(err, res));
 };
